+++ dotnet-roslyn/analyzers/Audacia.CodeAnalysis.Analyzers/CHANGELOG.md
@@ -6,7 +6,7 @@
 
 ### Changed
 - Updated ACL1020 ("AssertionRequired") so mock verification calls are treated as assertions.
-    - Added support for `Moq` verification methods (`Verify*`).
+    - Added support for `Moq` verification methods (`Verify`, `VerifyAll`, `VerifyGet`, `VerifySet`, `VerifyAdd`, `VerifyRemove`, and `VerifyNoOtherCalls`).
     - Added support for `NSubstitute` verification methods (`Received*`, `DidNotReceive*`, and `InOrder`).
 
 # 1.13.0 - 2026-06-04